Use ex17 to write a string editing script
sing exercise 17, I wrote a script that would write an old file
to a new file based on filenames added in the terminal.
I took this a step further and asked for user input that
would be added to what was written in the original file and
put into the new file.